+++ cloud/local/process.go
@@ -9,7 +9,7 @@ import (
 	sched "github.com/datacol-io/datacol/k8s"
 )
 
-func (g *LocalCloud) ProcessRun(name string, stream io.ReadWriter, command []string) error {
+func (g *LocalCloud) ProcessRun(name string, stream io.ReadWriter, opts pb.ProcessRunOptions) error {
 	ns := g.Name
 	cfg, err := getKubeClientConfig(ns)
 	if err != nil {
@@ -19,7 +19,7 @@ func (g *LocalCloud) ProcessRun(name string, stream io.ReadWriter, command []str
 	app, _ := g.AppGet(name)
 	envVars, _ := g.EnvironmentGet(name)
 
-	return sched.ProcessRun(g.kubeClient(), cfg, ns, name, g.latestImage(app), command, envVars, false, stream, cloud.LocalProvider)
+	return sched.ProcessRun(g.kubeClient(), cfg, ns, name, g.latestImage(app), opts, envVars, false, stream, cloud.LocalProvider)
 }
 
 func (g *LocalCloud) ProcessList(app string) ([]*pb.Process, error) {